requirements.txt
flask
flask-cors
azure-identity
azure-core
azure-ai-projects
azure-ai-agents
pytz
python-dotenv